+++ src/main/java/online/bottler/slack/application/port/in/SlackUseCase.java
@@ -1,0 +1,9 @@
+package online.bottler.slack.application.port.in;
+
+import online.bottler.slack.domain.SlackConstant;
+
+public interface SlackUseCase {
+
+    void sendSlackMessage(SlackConstant slackConstant, Long userId);
+
+}
